requirements.txt
sentence-transformers==2.2.2
transformers==4.35.0
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-dotenv==1.0.0
pyyaml==6.0
numpy==1.24.3
pandas==2.0.3
tqdm==4.66.1
pytest==7.4.0